Conflicts:
	cli/src/main/java/co/cask/cdap/shell/CLIMain.java
	cli/src/main/java/co/cask/cdap/shell/ProgramIdCompleterFactory.java
	cli/src/main/java/co/cask/cdap/shell/command/CallCommandSet.java
	cli/src/main/java/co/cask/cdap/shell/command/CallProcedureCommand.java
	cli/src/main/java/co/cask/cdap/shell/command/CommandMatch.java
	cli/src/main/java/co/cask/cdap/shell/command/ConnectCommand.java
	cli/src/main/java/co/cask/cdap/shell/command/CreateCommandSet.java
	cli/src/main/java/co/cask/cdap/shell/command/CreateDatasetCommandSet.java
	cli/src/main/java/co/cask/cdap/shell/command/CreateDatasetInstanceCommand.java
	cli/src/main/java/co/cask/cdap/shell/command/CreateStreamCommand.java
	cli/src/main/java/co/cask/cdap/shell/command/DeleteAppCommand.java
	cli/src/main/java/co/cask/cdap/shell/command/DeleteCommandSet.java
	cli/src/main/java/co/cask/cdap/shell/command/DeleteDatasetCommandSet.java
	cli/src/main/java/co/cask/cdap/shell/command/DeleteDatasetInstanceCommand.java
	cli/src/main/java/co/cask/cdap/shell/command/DeleteDatasetModuleCommand.java
	cli/src/main/java/co/cask/cdap/shell/command/DeployAppCommand.java
	cli/src/main/java/co/cask/cdap/shell/command/DeployCommandSet.java
	cli/src/main/java/co/cask/cdap/shell/command/DeployDatasetCommandSet.java
	cli/src/main/java/co/cask/cdap/shell/command/DeployDatasetModuleCommand.java
	cli/src/main/java/co/cask/cdap/shell/command/DescribeAppCommand.java
	cli/src/main/java/co/cask/cdap/shell/command/DescribeCommandSet.java
	cli/src/main/java/co/cask/cdap/shell/command/DescribeDatasetCommandSet.java
	cli/src/main/java/co/cask/cdap/shell/command/DescribeDatasetModuleCommand.java
	cli/src/main/java/co/cask/cdap/shell/command/DescribeDatasetTypeCommand.java
	cli/src/main/java/co/cask/cdap/shell/command/DescribeStreamCommand.java
	cli/src/main/java/co/cask/cdap/shell/command/ExecuteQueryCommand.java
	cli/src/main/java/co/cask/cdap/shell/command/GetCommandSet.java
	cli/src/main/java/co/cask/cdap/shell/command/GetHistoryCommandSet.java
	cli/src/main/java/co/cask/cdap/shell/command/GetInstancesCommandSet.java
	cli/src/main/java/co/cask/cdap/shell/command/GetLiveInfoCommandSet.java
	cli/src/main/java/co/cask/cdap/shell/command/GetLogsCommandSet.java
	cli/src/main/java/co/cask/cdap/shell/command/GetProgramHistoryCommand.java
	cli/src/main/java/co/cask/cdap/shell/command/GetProgramInstancesCommand.java
	cli/src/main/java/co/cask/cdap/shell/command/GetProgramLiveInfoCommand.java
	cli/src/main/java/co/cask/cdap/shell/command/GetProgramLogsCommand.java
	cli/src/main/java/co/cask/cdap/shell/command/GetProgramStatusCommand.java
	cli/src/main/java/co/cask/cdap/shell/command/GetStatusCommandSet.java
	cli/src/main/java/co/cask/cdap/shell/command/GetStreamEventsCommand.java
	cli/src/main/java/co/cask/cdap/shell/command/ListAllProgramsCommand.java
	cli/src/main/java/co/cask/cdap/shell/command/ListAppsCommand.java
	cli/src/main/java/co/cask/cdap/shell/command/ListCommandSet.java
	cli/src/main/java/co/cask/cdap/shell/command/ListDatasetCommandSet.java
	cli/src/main/java/co/cask/cdap/shell/command/ListDatasetInstancesCommand.java
	cli/src/main/java/co/cask/cdap/shell/command/ListDatasetModulesCommand.java
	cli/src/main/java/co/cask/cdap/shell/command/ListDatasetTypesCommand.java
	cli/src/main/java/co/cask/cdap/shell/command/ListProgramsCommand.java
	cli/src/main/java/co/cask/cdap/shell/command/ListStreamsCommand.java
	cli/src/main/java/co/cask/cdap/shell/command/SendCommandSet.java
	cli/src/main/java/co/cask/cdap/shell/command/SendStreamEventCommand.java
	cli/src/main/java/co/cask/cdap/shell/command/SetCommandSet.java
	cli/src/main/java/co/cask/cdap/shell/command/SetInstancesCommandSet.java
	cli/src/main/java/co/cask/cdap/shell/command/SetProgramInstancesCommand.java
	cli/src/main/java/co/cask/cdap/shell/command/SetStreamCommandSet.java
	cli/src/main/java/co/cask/cdap/shell/command/SetStreamTTLCommand.java
	cli/src/main/java/co/cask/cdap/shell/command/StartProgramCommand.java
	cli/src/main/java/co/cask/cdap/shell/command/StartProgramCommandSet.java
	cli/src/main/java/co/cask/cdap/shell/command/StopProgramCommand.java
	cli/src/main/java/co/cask/cdap/shell/command/StopProgramCommandSet.java
	cli/src/main/java/co/cask/cdap/shell/command/TruncateCommandSet.java
	cli/src/main/java/co/cask/cdap/shell/command/TruncateDatasetCommandSet.java
	cli/src/main/java/co/cask/cdap/shell/command/TruncateDatasetInstanceCommand.java
	cli/src/main/java/co/cask/cdap/shell/command/TruncateStreamCommand.java
	cli/src/main/java/co/cask/cdap/shell/command/common/ExitCommand.java
	cli/src/main/java/co/cask/cdap/shell/command/create/CreateCommandSet.java
	cli/src/main/java/co/cask/cdap/shell/command/set/SetInstancesCommandSet.java
	cli/src/main/java/co/cask/cdap/shell/command/set/SetProgramInstancesCommandSet.java

+++ app-fabric/src/main/java/co/cask/cdap/test/RuntimeStats.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2014 Cask Data, Inc.
+ * Copyright © 2014 Cask Data, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy of
@@ -33,7 +33,7 @@ public final class RuntimeStats {
 
   private static ConcurrentMap<String, AtomicLong> counters = Maps.newConcurrentMap();
 
-  public static void count(String name, int count) {
+  public static void count(String name, long count) {
     AtomicLong oldValue = counters.putIfAbsent(name, new AtomicLong(count));
     if (oldValue != null) {
       oldValue.addAndGet(count);
