+++ watchdog/src/main/java/co/cask/cdap/metrics/data/AggregatesTable.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2014 Cask Data, Inc.
+ * Copyright © 2014 Cask Data, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy of
@@ -103,11 +103,11 @@ public final class AggregatesTable {
         Map<byte[], Long> increments = Maps.newTreeMap(Bytes.BYTES_COMPARATOR);
 
         // The no tag value
-        increments.put(Bytes.toBytes(MetricsConstants.EMPTY_TAG), (long) record.getValue());
+        increments.put(Bytes.toBytes(MetricsConstants.EMPTY_TAG), record.getValue());
 
         // For each tag, increments corresponding values
         for (TagMetric tag : record.getTags()) {
-          increments.put(Bytes.toBytes(tag.getTag()), (long) tag.getValue());
+          increments.put(Bytes.toBytes(tag.getTag()), tag.getValue());
         }
         aggregatesTable.increment(rowKey, increments);
       }
