+++ watchdog/src/main/java/co/cask/cdap/metrics/data/TimeValue.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2014 Cask Data, Inc.
+ * Copyright © 2014 Cask Data, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy of
@@ -21,9 +21,9 @@ package co.cask.cdap.metrics.data;
 public final class TimeValue {
 
   private final long time;
-  private final int value;
+  private final long value;
 
-  public TimeValue(long time, int value) {
+  public TimeValue(long time, long value) {
     this.time = time;
     this.value = value;
   }
@@ -32,7 +32,7 @@ public final class TimeValue {
     return time;
   }
 
-  public int getValue() {
+  public long getValue() {
     return value;
   }
 }
