+++ watchdog/src/main/java/co/cask/cdap/metrics/data/TimeValueIterable.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2014 Cask Data, Inc.
+ * Copyright © 2014 Cask Data, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy of
@@ -59,7 +59,7 @@ final class TimeValueIterable implements Iterable<TimeValue> {
           if (timestamp > endTime) {
             return endOfData();
           }
-          return new TimeValue(timestamp, Bytes.toInt(entry.getValue()));
+          return new TimeValue(timestamp, Bytes.toLong(entry.getValue()));
         }
         return endOfData();
       }
