+++ watchdog/src/main/java/co/cask/cdap/metrics/query/MetricsRequestExecutor.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2014 Cask Data, Inc.
+ * Copyright © 2014 Cask Data, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy of
@@ -160,8 +160,8 @@ public class MetricsRequestExecutor {
 
     for (int i = 0; i < metricsRequest.getCount(); i++) {
       long resultTime = metricsRequest.getStartTime() + i;
-      int tupleRead = 0;
-      int eventProcessed = 0;
+      long tupleRead = 0;
+      long eventProcessed = 0;
       if (tuplesReadItor.hasNext() && tuplesReadItor.peek().getTime() == resultTime) {
         tupleRead = tuplesReadItor.next().getValue();
       }
